feat!: move babel options to task options and allow configuring env
