+++ lib/schemas/src/tasks/babel.ts
@@ -3,8 +3,10 @@ import { z } from 'zod'
 export const BabelSchema = z.object({
   files: z.string().optional(),
   outputPath: z.string().optional(),
-  configFile: z.string().optional()
+  configFile: z.string().optional(),
+  envName: z.union([z.literal('production'), z.literal('development')])
 })
+
 export type BabelOptions = z.infer<typeof BabelSchema>
 
 export const Schema = BabelSchema
