+++ plugins/babel/.toolkitrc.yml
@@ -1,8 +1,7 @@
 tasks:
-  BabelDevelopment: './lib/tasks/development'
-  BabelProduction: './lib/tasks/production'
+  Babel: './lib/tasks/babel'
 
 commands:
-  'build:local': BabelDevelopment
-  'build:ci': BabelProduction
-  'build:remote': BabelProduction
+  'build:local': Babel
+  'build:ci': Babel
+  'build:remote': Babel
